test: unify all pallets test to use block_no as u32

+++ pallets/contracts/src/wasm/mod.rs
@@ -544,7 +544,7 @@ mod tests {
             self.events.push((topics, data))
         }
 
-        fn block_number(&self) -> u64 {
+        fn block_number(&self) -> u32 {
             121
         }
